Remove MoveRequest and add WalkableTileDTO support
This commit removes the `MoveRequest` class from the API, indicating a shift in functionality. A new property `WalkableTiles` has been added to the `GameLoadDTO` class to include information about walkable tiles in the game state. The `MazeGameService` has been updated to populate this new property with walkable tile data, mapping them to instances of the newly created `WalkableTileDTO` class, which contains X and Y coordinates.